Feat(rename):
Rename ConfigLoader to JestConfigLoader
Rename DefaultConfigLoader to DefaultJestConfigLoader
Rename ReactConfigLoader to ReactScriptsJestConfigLoader
