+++ src/configLoaders/ReactScriptsJestConfigLoader.ts
